# swagger.json
{
	"components": {
...
		"schemas": {
...
			"IMessage": {
				"description": "Respuesta simple con mensaje.",
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"message"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"status": 404,
					"message": "NOT FOUND"
				}
			},
...
			"User": {
				"description": "Esquema con el cual el backend retorna un usuario.",
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"passwordHash": {
						"type": "string"
					},
					"phone": {
						"type": "number",
						"format": "double"
					},
					"birthday": {
						"type": "string",
						"format": "date-time"
					},
					"disabled": {
						"type": "boolean"
					},
					"role": {
						"properties": {
							"name": {
								"type": "string"
							},
							"_id": {
								"type": "string"
							}
						},
						"required": [
							"name",
							"_id"
						],
						"type": "object"
					},
					"_v": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"name",
					"lastName",
					"email",
					"passwordHash",
					"phone",
					"birthday",
					"disabled",
					"role",
					"_v"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"_id": "6531618e6025da22956875a6",
					"name": "user",
					"lastName": "admin",
					"email": "user@email.com",
					"passwordHash": "$argon2id$v=19$m=65536,t=3,p=4$syRur4yCeHRcZe6bp0KfRA$fLli/g+d79zcs2M/MCWLGgj2JRPWb3wjTaUhPORdWxI",
					"phone": 10000000000,
					"birthday": "1995-10-20T04:03:42.164Z",
					"role": {
						"_id": "65305445746510934b074e05",
						"name": "Administrador"
					},
					"disabled": false,
					"_v": 0
				}
			},
...
			"UserCreation": {
				"description": "Esquema para crear un usuario.",
				"properties": {
					"name": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phone": {
						"type": "number",
						"format": "double"
					},
					"birthday": {
						"type": "string",
						"format": "date-time"
					},
					"role": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"lastName",
					"email",
					"password",
					"phone",
					"birthday",
					"role"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"name": "pepito",
					"lastName": "perez",
					"email": "elpepe@email.com",
					"password": "0123456789ABCDEF",
					"phone": 10000000001,
					"birthday": "2000-10-20T04:03:42.164Z",
					"role": "65305445746510934b074e05"
				}
			},
			"UserUpdate": {
				"description": "Esquema para editar un usuario.",
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"lastName": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phone": {
						"type": "number",
						"format": "double"
					},
					"role": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false,
				"example": {
					"_id": "6531618e6025da22956875a6",
					"name": "pepito",
					"lastName": "perez",
					"email": "elpepe@email.com",
					"password": "0123456789ABCDEF",
					"phone": 10000000001,
					"role": "65305445746510934b074e05"
				}
			}
		},
		"securitySchemes": {
			"token": {
				"type": "http",
				"description": "JWT token based security",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"scopes": {
					"create": "Create a resource",
					"update": "Update a resource",
					"delete": "Delete a resource"
				}
			}
		}
	},
	"info": {
		"title": "Proyecto 1 Backend",
		"version": "1.0.0",
		"description": "API para ofrecer servicios de domicilios de restaurantes.",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "david"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Obtiene todos los usuarios.",
				"summary": "Get Users",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"description": "Registra un nuevo usuario en la base de datos.",
				"summary": "Create User",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "El usuario a registrar",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreation",
								"description": "El usuario a registrar"
							}
						}
					}
				}
			},
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"description": "Edita los datos de un usuario. El usuario es identificado por la propiedad _id.",
				"summary": "Update User",
				"tags": [
					"Users"
				],
				"security": [
					{
						"token": [
							"update"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "El usuario a editar. Todos los campos que estén definidos en este objeto se sobreescriben.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdate",
								"description": "El usuario a editar. Todos los campos que estén definidos en este objeto se sobreescriben."
							}
						}
					}
				}
			}
		},
...
		"/users/{id}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				},
				"description": "Obtiene los datos de un usuario.",
				"summary": "Get User",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"description": "Id del usuario",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMessage"
								}
							}
						}
					}
				},
				"description": "Deshabilita un usuario, este ya no se podrá leer.",
				"summary": "Delete User",
				"tags": [
					"Users"
				],
				"security": [
					{
						"token": [
							"delete"
						]
					}
				],
				"parameters": [
					{
						"description": "Id del usuario.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}